获取新添加的主键

+++ parasol-file/parasol-file-api/src/main/java/com/ginkgocap/parasol/file/service/UserFileCategoryServer.java
@@ -15,7 +15,7 @@ public interface UserFileCategoryServer {
      * @param pojo
      * @return
      */
-    int insert(UserFileCategory pojo);
+    long insert(UserFileCategory pojo);
 
     /**
      * 集合方式添加用户文件或者目录
