+++ parasol-file/parasol-file-provider/src/main/java/com/ginkgocap/parasol/file/dao/UserFileCategoryDao.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Component;
 @Component
 public interface UserFileCategoryDao {
 
-    int insert(UserFileCategory pojo);
+    long insert(UserFileCategory pojo);
 
     int insertList(@Param("pojos") List< UserFileCategory> pojo);
